+++ borsh/src/lib.rs
@@ -12,9 +12,10 @@ pub mod schema;
 pub mod schema_helpers;
 pub mod ser;
 
+pub use de::from_slice;
 pub use de::BorshDeserialize;
 pub use schema::BorshSchema;
-pub use schema_helpers::{from_slice, try_from_slice_with_schema, try_to_vec_with_schema};
+pub use schema_helpers::{try_from_slice_with_schema, try_to_vec_with_schema};
 pub use ser::helpers::{to_vec, to_writer};
 pub use ser::BorshSerialize;
